requirements.txt
python-dotenv~=1.0.1
pymilvus~=2.6.6
transformers~=4.57.3
torch~=2.3.1+cu118
langchain-core~=1.2.5
langchain-experimental~=0.3.4
huggingface-hub~=0.36.0
html2text~=2025.4.15
requests~=2.32.5
beautifulsoup4~=4.12.3
urllib3~=2.6.2
aiohttp~=3.11.11
numpy~=1.26.4
python-docx~=1.1.2
langchain-openai~=1.1.6
langgraph~=1.0.5
mcp~=1.25.0
psycopg~=3.3.2